Add interactive rotation feature
Hold control-shift while dragging near the right edge, and it will
rotate the image continuously while you drag up and down.  On release,
it transforms the image in memory.  Then the rotation can be saved to
the file, as with other transformations.

Rotation must be stored in floating point: a one-degree increment is
not nearly accurate enough for rotating photos that were taken or
scanned crooked.  QTransform::rotate() takes it as qreal anyhow.

Since QWidget (and QLabel) does not have an arbitrary rotation
feature (unlike QGraphicsItem and QQuickItem), we need to have
the actual image widget larger than the image so that rendering
doesn't get cut off while the image is being rotated.  So it can't
be a QLabel anymore; it needs to be a custom widget that can call
QPainter::rotate() before painting the image.

Warning: this patch introduces problems with zooming and scrolling.
It works well when the window is larger than the image being shown
so that scaling is 100% and there is space around the edges to show
the rotation.  But just as ImageWidget has a rotation property,
it probably needs to have a scale property too, to handle the
control-mousewheel zooming properly.  And of course it needs to
have bigger geometry when you zoom, but I think that's done outside
so far.