Store average brightness in ThumbsViewer's model; show in Image Info
This will help to implement the feature to select by brightness,
and help the user decide how bright is too bright or too dark.
The method chosen is to scale each thumbnail to a 1x1 pixel image,
and take the greyscale value of that pixel.  Brightness is in the
range 0..1 (QImage uses 0..255 but that's not very future-proof,
and users probably prefer to think in terms of percentage anyway).