Add feature: select images by brightness range
When doing batch processing, it's nice to be able to find all
the images which are too bright or too dark to be useful.
After selecting them, it's easy to delete them.

We need to know the brightnesses of all the images to do this.
The way chosen here is to preload all their thumbnails.
But that can take some time, so we show a progress dialog, with
the ability to cancel.  The dialog is not modal, because it appears
that loading thumbnails does not interfere with other usage of the
application, so it can be done in the background.