Store tags in xattrs; allow turning off writing of EXIF keywords
I don't like modifying my original photos, because writing of any archival
file is dangerous, and because it changes the timestamp of the file.

A suitable alternative for writing tags is to use the standard that KDE
supports: write to the extended attribute user.xdg.tags.  This is what
Dolphin does for example when one opens a file's Properties | Details and
edits the tags: https://vhanda.in/blog/2014/07/tagging-your-files/
However that requires multiple steps per file, and now phototonic makes it
much easier.

Added a preference to enable/disable writing the EXIF keywords, since
that's redundant if one is using the xattr mechanism instead.  But the
EXIF keywords are always read and shown as tags; thus when viewing any
photo, if one then edits the tags, the whole set of tags should be written
to the extended attribute.